Update patch set 1
Patch Set 1:

You might want to drop the Fraser prefix in these documents and in general. For each release, we will just modify the document.

Also, I believe this name will show up in the overall URL path for the toplevel OPNFV docs. See the examples below for Yardstick and Functest. If the name stays consistent, someone can change the 'stable-<release>' portion of the URL for each release.

http://docs.opnfv.org/en/stable-euphrates/submodules/yardstick/docs/release/release-notes/release-notes.html
http://docs.opnfv.org/en/stable-fraser/submodules/yardstick/docs/release/release-notes/release-notes.html

http://docs.opnfv.org/en/stable-euphrates/submodules/functest/docs/release/release-notes/functest-release.html
http://docs.opnfv.org/en/stable-fraser/submodules/functest/docs/release/release-notes/functest-release.html

Patch-set: 1
Label: Code-Review=0